Fixed spring bean creation not working on inner class
