+++ src/main/java/uk/gov/companieshouse/efs/web/controller/DocumentUploadControllerImpl.java
@@ -238,35 +238,4 @@ public class DocumentUploadControllerImpl extends BaseControllerImpl implements 
     }
 
 
-    /**
-     * Parameter object used to group file transfer related objects.
-     * Reduces the number of parameters taken by the constructor and makes sonar happy.
-     */
-    static class FileTransferManager {
-        private final FileTransferApiClient fileTransferApiClient;
-        private final FileUploadConfiguration fileUploadConfiguration;
-        private final DocumentUploadValidator documentUploadValidator;
-
-        @Autowired
-        private FileTransferManager(FileTransferApiClient fileTransferApiClient,
-                                    FileUploadConfiguration fileUploadConfiguration,
-                                    DocumentUploadValidator documentUploadValidator) {
-
-            this.fileTransferApiClient = fileTransferApiClient;
-            this.fileUploadConfiguration = fileUploadConfiguration;
-            this.documentUploadValidator = documentUploadValidator;
-        }
-
-        public FileTransferApiClient getFileTransferApiClient() {
-            return fileTransferApiClient;
-        }
-
-        public FileUploadConfiguration getFileUploadConfiguration() {
-            return fileUploadConfiguration;
-        }
-
-        public DocumentUploadValidator getDocumentUploadValidator() {
-            return documentUploadValidator;
-        }
-    }
 }
