+++ src/test/java/uk/gov/companieshouse/efs/web/controller/DocumentUploadControllerTest.java
@@ -65,7 +65,7 @@ class DocumentUploadControllerTest extends BaseControllerImplTest {
     private ModelMap attributes;
 
     @Mock
-    DocumentUploadControllerImpl.FileTransferManager paramObject;
+    FileTransferManager paramObject;
 
     @InjectMocks
     private DocumentUploadControllerImpl toTest;
